+++ test/test_browser.py
@@ -418,7 +418,7 @@ If manually bisecting:
 
   # Tests that user .html shell files can manually download .data files created with --preload-file cmdline.
   @parameterized({
-    'default': ([],),
+    '': ([],),
     'pthreads': (['-pthread', '-sPROXY_TO_PTHREAD', '-sEXIT_RUNTIME'],),
   })
   @requires_threads
@@ -5499,7 +5499,8 @@ Module["preRun"].push(function () {
 
   # Tests the AudioWorklet demo
   @parameterized({
-    'default': ([],),
+    '': ([],),
+    'memory64': (['-sMEMORY64', '-Wno-experimental'],),
     'with_fs': (['--preload-file', test_file('hello_world.c') + '@/'],),
     'closure': (['--closure', '1', '-Oz'],),
     'asyncify': (['-sASYNCIFY'],),
@@ -5509,11 +5510,13 @@ Module["preRun"].push(function () {
     'minimal_runtime_pthreads_and_closure': (['-sMINIMAL_RUNTIME', '-pthread', '--closure', '1', '-Oz'],),
   })
   def test_audio_worklet(self, args):
+    if '-sMEMORY64' in args and is_firefox():
+      self.skipTest('https://github.com/emscripten-core/emscripten/issues/19161')
     self.btest_exit(test_file('webaudio/audioworklet.c'), args=['-sAUDIO_WORKLET', '-sWASM_WORKERS'] + args)
 
   # Tests that posting functions between the main thread and the audioworklet thread works
   @parameterized({
-    'default': ([],),
+    '': ([],),
     'closure': (['--closure', '1', '-Oz'],),
   })
   def test_audio_worklet_post_function(self, args):