+++ ChangeLog.md
@@ -23,6 +23,8 @@ See docs/process.md for more on how version tagging works.
 - Enable WASM_BIGINT support by default. (#19156)
 - Bump the default minimum Firefox version from 65 to 68.
 - Bump the default minimum Node version from 10.19 to 15.0.
+- The `USES_DYNAMIC_ALLOC` setting has been deprecated.  You can get the same
+  effect from `-sMALLOC=none`. (#19164)
 
 3.1.35 - 04/03/23
 -----------------